+++ conda_recipe_manager/fetcher/api/__init__.py
@@ -1,0 +1,3 @@
+"""
+:Description: Module that provides API models for remote resources.
+"""