+++ conda_recipe_manager/licenses/__init__.py
@@ -1,0 +1,3 @@
+"""
+:Description: Module that handles understanding and manipulating software licenses.
+"""