+++ conda_recipe_manager/scanner/dependency/__init__.py
@@ -1,0 +1,3 @@
+"""
+:Description: Module that provides tools for scanning/reading dependencies from various sources.
+"""