+++ conda_recipe_manager/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+:Description: General utility module for CRM components.
+"""