+++ CHANGELOG.md
@@ -9,7 +9,12 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 
 ## [Unreleased]
 ### Added
+- Logging support via the standard Python logging library. By default, all modules (except `commands`,
+  which contains the CLI programs) default to use the `NullHandler` so that the client program can determine
+  the best way to log for their needs.
 ### Changed
+- De-emphasized/redacted the `MessageTable` class in favor of the standard Python library logger. It is now only
+  used as part of the recipe conversion process and `crm convert`
 ### Deprecated
 ### Removed
 ### Fixed